+++ src/com/grivera/generator/sensors/StorageNode.java
@@ -9,27 +9,27 @@ public class StorageNode extends SensorNode {
 
     private static final double E_store = 100e-9;
 
-    private static int idCounter = 1;
-    private int id;
-    private int capacity;
-    private int usedSpace;
+    private static long idCounter = 1;
+    private long id;
+    private long capacity;
+    private long usedSpace;
 
-    public StorageNode(double x, double y, double tr, int capacity) {
+    public StorageNode(double x, double y, double tr, long capacity) {
         super(x, y, tr, String.format("SN%02d", idCounter));
         this.id = idCounter++;
         this.setCapacity(capacity);
     }
 
-    public void setCapacity(int capacity) {
+    public void setCapacity(long capacity) {
         this.capacity = capacity;
         this.usedSpace = 0;
     }
 
-    public int getCapacity() {
+    public long getCapacity() {
         return this.capacity;
     }
 
-    public int getUsedSpace() {
+    public long getUsedSpace() {
         return this.usedSpace;
     }
 
@@ -37,11 +37,11 @@ public class StorageNode extends SensorNode {
         return this.usedSpace >= this.capacity;
     }
 
-    public boolean canStore(int deltaPackets) {
+    public boolean canStore(long deltaPackets) {
         return this.usedSpace + deltaPackets <= this.capacity;
     }
 
-    public void storePackets(int packets) {
+    public void storePackets(long packets) {
         if (!this.canStore(packets)) {
             throw new IllegalArgumentException(
                     String.format("%s cannot store %d packets (%d/%d full)",
@@ -58,11 +58,11 @@ public class StorageNode extends SensorNode {
     }
 
     @Override
-    public int getId() {
+    public long getId() {
         return this.id;
     }
 
-    public int getSpaceLeft() {
+    public long getSpaceLeft() {
         return this.capacity - this.usedSpace;
     }
